+++ packages/@ionic/cli-framework/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.4](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.1.3...@ionic/cli-framework@4.1.4) (2020-05-06)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+## [4.1.3](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.1.2...@ionic/cli-framework@4.1.3) (2020-04-29)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
+## [4.1.2](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.1.1...@ionic/cli-framework@4.1.2) (2020-03-30)
+
+**Note:** Version bump only for package @ionic/cli-framework
+
+
+
+
+
 ## [4.1.1](https://github.com/ionic-team/ionic-cli/compare/@ionic/cli-framework@4.1.0...@ionic/cli-framework@4.1.1) (2020-03-03)
 
 **Note:** Version bump only for package @ionic/cli-framework